Disable MSan in Swiftshader JIT
There are a number of tests that are (silently) failing with errors in
Swiftshader's JIT, e.g.

==218683==WARNING: MemorySanitizer: use-of-uninitialized-value
    #0 0x7f0d6b006d68 in ?? ??:0

SUMMARY: MemorySanitizer: use-of-uninitialized-value
(/memfd:swiftshader_jit (deleted)+0xd68)
  ORIGIN: invalid (0). Might be a bug in MemorySanitizer origin tracking.
    This could still be a bug in your code, too!

In an ideal world, there would be an investigation and a fix for these
reports. In practice, this is not happening as Swiftshader is no longer
actively maintained; in fact, Swiftshader is being removed as a fallback
for WebGL (crbug.com/40277080).

Per the LLVM documentation, omitting the llvm::Attribute::SanitizeMemory
annotation for a JIT function has the same effect as applying
` __attribute__((no_sanitize("memory")))` to C++ functions. This should
suppress the errors and unblock enabling more accurate MSan error
detection in tests.

Bug: 40267022
Change-Id: I3c201fcbae9b7563d327eb5ca358ebcbd242ea8a