Change enum values in IceIntrinsics.h to avoid implicit conversion
Signed values are stored using two's complement, so when a signed
value is converted to a 1-bit bitfield, it transforms the value 1
into -1, causing clang to issue a warning. See here for a more
detailed discussion: https://github.com/llvm/llvm-project/issues/53253
This warning is currently disabled in chrome (crbug.com/40234766),
but we'd like to re-enable it.

The conversion is happening with the enum values in IceIntrinsics.h.
To avoid it, we set the _T values to -1 instead of 1 at the beginning,
so there's no implicit conversion. This is rather ugly, since
intuitively we would like the "true" value to be 1, but it's the only
solution I'm aware of.

I tried several other options to force clang to treat the enum as an
unsigned value, but this causes it to change the memory layout of the
InstrinsicInfo struct, doubling its size (4 bytes -> 8 bytes).

Change-Id: I9bf0259bb81677b50951755118f2492300d166a4